Make the blockchain check the difficulty of new blocks

+++ src/util/config.rs
@@ -13,7 +13,7 @@ pub struct Config {
     // Miner settings
     pub max_blocks: u64,
     pub max_nonce: u64,
-    pub difficulty: usize,
+    pub difficulty: u32,
     pub tx_waiting_ms: u64,
 }
 
@@ -31,7 +31,7 @@ impl Config {
             // Miner settings
             max_blocks: Config::read_envvar::<u64>("MAX_BLOCKS", 0), // unlimited blocks
             max_nonce: Config::read_envvar::<u64>("MAX_NONCE", 1_000_000),
-            difficulty: Config::read_envvar::<usize>("DIFFICULTY", 10),
+            difficulty: Config::read_envvar::<u32>("DIFFICULTY", 10),
             tx_waiting_ms: Config::read_envvar::<u64>("TRANSACTION_WAITING_MS", 10000),
         }
     }
